+++ tests/expletive-standard.spec.ts
@@ -1,0 +1,58 @@
+// tslint:disable-next-line:import-name
+import dedent from 'ts-dedent';
+import { SpeechMarkdown } from '../src/SpeechMarkdown';
+
+describe('expletive-standard', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    You said (word)[bleep] and (word)[expletive] at school.
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      You said <say-as interpret-as="expletive">word</say-as> and <say-as interpret-as="expletive">word</say-as> at school.
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      You said <say-as interpret-as="expletive">word</say-as> and <say-as interpret-as="expletive">word</say-as> at school.
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      You said word and word at school.
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});