+++ tests/interjection-standard.spec.ts
@@ -1,0 +1,58 @@
+// tslint:disable-next-line:import-name
+import dedent from 'ts-dedent';
+import { SpeechMarkdown } from '../src/SpeechMarkdown';
+
+describe('interjection-standard', () => {
+
+  const speech = new SpeechMarkdown();
+
+  const markdown = dedent`
+    (Wow)[interjection], I didn't see that coming.
+  `;
+
+  test('converts to SSML - Amazon Alexa', () => {
+
+    const options = {
+      platform: 'amazon-alexa'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      <say-as interpret-as="interjection">Wow</say-as>, I didn't see that coming.
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to SSML - Google Assistant', () => {
+
+    const options = {
+      platform: 'google-assistant'
+    };
+    const ssml = speech.toSSML(markdown, options);
+
+    const expected = dedent`
+      <speak>
+      Wow, I didn't see that coming.
+      </speak>
+    `;
+
+    expect(ssml).toBe(expected);
+  });
+
+  test('converts to Plain Text', () => {
+
+    const options = {
+    };
+    const text = speech.toText(markdown, options);
+
+    const expected = dedent`
+      Wow, I didn't see that coming.
+    `;
+
+    expect(text).toBe(expected);
+  });
+
+});